GAWB-3786 extracted bond provider logic out of martha_v2.js into common/helpers.js so it can be used by fileSummaryV1.js
